features: break out feature_state into its own header
